Move ephemeral clients to _core modules
