+++ src/vercel/projects/__init__.py
@@ -7,28 +7,9 @@ from typing import Any
 from .._http import (
     DEFAULT_API_BASE_URL,
     DEFAULT_TIMEOUT,
-    BlockingTransport,
-    HTTPConfig,
     iter_coroutine,
 )
-from ._core import _BaseProjectsClient
-
-
-class _EphemeralSyncClient(_BaseProjectsClient):
-    """Internal sync client for module-level functions."""
-
-    def __init__(
-        self,
-        token: str | None,
-        base_url: str,
-        timeout: float,
-    ) -> None:
-        config = HTTPConfig(
-            base_url=base_url,
-            timeout=timeout,
-            token=token,
-        )
-        self._transport = BlockingTransport(config)
+from ._core import SyncProjectsClient
 
 
 def get_projects(
@@ -52,7 +33,7 @@ def get_projects(
 
     Returns: dict with keys like {"projects": [...], "pagination": {...}}
     """
-    client = _EphemeralSyncClient(token, base_url, timeout)
+    client = SyncProjectsClient(token, base_url, timeout)
     return iter_coroutine(
         client._get_projects(team_id=team_id, slug=slug, query=query)
     )
@@ -72,7 +53,7 @@ def create_project(
     body: JSON payload (must include at least name)
     Optional query params: team_id -> teamId, slug -> slug
     """
-    client = _EphemeralSyncClient(token, base_url, timeout)
+    client = SyncProjectsClient(token, base_url, timeout)
     return iter_coroutine(
         client._create_project(body=body, team_id=team_id, slug=slug)
     )
@@ -89,7 +70,7 @@ def update_project(
     timeout: float = DEFAULT_TIMEOUT,
 ) -> dict[str, Any]:
     """Update an existing project by id or name."""
-    client = _EphemeralSyncClient(token, base_url, timeout)
+    client = SyncProjectsClient(token, base_url, timeout)
     return iter_coroutine(
         client._update_project(id_or_name, body=body, team_id=team_id, slug=slug)
     )
@@ -105,7 +86,7 @@ def delete_project(
     timeout: float = DEFAULT_TIMEOUT,
 ) -> None:
     """Delete a project by id or name. Returns None on success (204)."""
-    client = _EphemeralSyncClient(token, base_url, timeout)
+    client = SyncProjectsClient(token, base_url, timeout)
     return iter_coroutine(
         client._delete_project(id_or_name, team_id=team_id, slug=slug)
     )
