+++ src/vercel/projects/aio.py
@@ -7,27 +7,8 @@ from typing import Any
 from .._http import (
     DEFAULT_API_BASE_URL,
     DEFAULT_TIMEOUT,
-    AsyncTransport,
-    HTTPConfig,
 )
-from ._core import _BaseProjectsClient
-
-
-class _EphemeralAsyncClient(_BaseProjectsClient):
-    """Internal async client for module-level functions."""
-
-    def __init__(
-        self,
-        token: str | None,
-        base_url: str,
-        timeout: float,
-    ) -> None:
-        config = HTTPConfig(
-            base_url=base_url,
-            timeout=timeout,
-            token=token,
-        )
-        self._transport = AsyncTransport(config)
+from ._core import AsyncProjectsClient
 
 
 async def get_projects(
@@ -51,7 +32,7 @@ async def get_projects(
 
     Returns: dict with keys like {"projects": [...], "pagination": {...}}
     """
-    client = _EphemeralAsyncClient(token, base_url, timeout)
+    client = AsyncProjectsClient(token, base_url, timeout)
     return await client._get_projects(team_id=team_id, slug=slug, query=query)
 
 
@@ -69,7 +50,7 @@ async def create_project(
     body: JSON payload (must include at least name)
     Optional query params: team_id -> teamId, slug -> slug
     """
-    client = _EphemeralAsyncClient(token, base_url, timeout)
+    client = AsyncProjectsClient(token, base_url, timeout)
     return await client._create_project(body=body, team_id=team_id, slug=slug)
 
 
@@ -84,7 +65,7 @@ async def update_project(
     timeout: float = DEFAULT_TIMEOUT,
 ) -> dict[str, Any]:
     """Update an existing project by id or name."""
-    client = _EphemeralAsyncClient(token, base_url, timeout)
+    client = AsyncProjectsClient(token, base_url, timeout)
     return await client._update_project(id_or_name, body=body, team_id=team_id, slug=slug)
 
 
@@ -98,7 +79,7 @@ async def delete_project(
     timeout: float = DEFAULT_TIMEOUT,
 ) -> None:
     """Delete a project by id or name. Returns None on success (204)."""
-    client = _EphemeralAsyncClient(token, base_url, timeout)
+    client = AsyncProjectsClient(token, base_url, timeout)
     return await client._delete_project(id_or_name, team_id=team_id, slug=slug)
